@@ -9,6 +9,7 @@ import com.google.common.collect.HashMultimap;
 import com.google.common.collect.Multimap;
 import com.google.common.collect.Sets;
 
+import dev.emi.trinkets.TrinketModifiers;
 import dev.emi.trinkets.api.SlotAttributes;
 import dev.emi.trinkets.api.SlotReference;
 import dev.emi.trinkets.api.SlotType;
@@ -16,6 +17,8 @@ import dev.emi.trinkets.api.Trinket;
 import dev.emi.trinkets.api.TrinketComponent;
 import dev.emi.trinkets.api.TrinketInventory;
 import dev.emi.trinkets.api.TrinketsApi;
+import dev.emi.trinkets.api.event.TrinketEquipCallback;
+import dev.emi.trinkets.api.event.TrinketUnequipCallback;
 import dev.emi.trinkets.payload.SyncInventoryPayload;
 import net.minecraft.component.EnchantmentEffectComponentTypes;
 import net.minecraft.entity.attribute.EntityAttributeInstance;
@@ -57,9 +60,9 @@ import net.minecraft.world.World;
 public abstract class LivingEntityMixin extends Entity {
 	@Unique
 	private final Map<String, ItemStack> lastEquippedTrinkets = new HashMap<>();
-	
+
 	@Shadow
-	protected abstract AttributeContainer getAttributes();
+	public abstract AttributeContainer getAttributes();
 
 	private LivingEntityMixin() {
 		super(null, null);
@@ -91,7 +94,7 @@ public abstract class LivingEntityMixin extends Entity {
 			DropRule dropRule = TrinketsApi.getTrinket(stack.getItem()).getDropRule(stack, ref, entity);
 
 			dropRule = TrinketDropCallback.EVENT.invoker().drop(dropRule, stack, ref, entity);
-			
+
 			TrinketInventory inventory = ref.inventory();
 
 			if (dropRule == DropRule.DEFAULT) {
@@ -123,15 +126,13 @@ public abstract class LivingEntityMixin extends Entity {
 		}));
 	}
 
+	@Unique
 	private void dropFromEntity(ItemStack stack) {
-		ItemEntity entity = dropStack(stack);
 		// Mimic player drop behavior for only players
-		if (entity != null && ((Entity) this) instanceof PlayerEntity) {
-			entity.setPos(entity.getX(), this.getEyeY() - 0.3, entity.getZ());
-			entity.setPickupDelay(40);
-			float magnitude = this.random.nextFloat() * 0.5f;
-			float angle = this.random.nextFloat() * ((float)Math.PI * 2);
-			entity.setVelocity(-MathHelper.sin(angle) * magnitude, 0.2f, MathHelper.cos(angle) * magnitude);
+		if (((Entity) this) instanceof PlayerEntity player) {
+			ItemEntity entity = player.dropItem(stack, true, false);
+		} else {
+			ItemEntity entity = dropStack(stack);
 		}
 	}
 
@@ -156,16 +157,16 @@ public abstract class LivingEntityMixin extends Entity {
 				if (!ItemStack.areEqual(newStack, oldStack)) {
 
 					TrinketsApi.getTrinket(oldStack.getItem()).onUnequip(oldStack, ref, entity);
+					TrinketUnequipCallback.EVENT.invoker().onUnequip(oldStack, ref, entity);
 					TrinketsApi.getTrinket(newStack.getItem()).onEquip(newStack, ref, entity);
+					TrinketEquipCallback.EVENT.invoker().onEquip(newStack, ref, entity);
 
 					World world = this.getWorld();
 					if (!world.isClient) {
 						contentUpdates.put(newRef, newStackCopy);
-						Identifier identifier = SlotAttributes.getIdentifier(ref);
 
 						if (!oldStack.isEmpty()) {
-							Trinket trinket = TrinketsApi.getTrinket(oldStack.getItem());
-							Multimap<RegistryEntry<EntityAttribute>, EntityAttributeModifier> map = trinket.getModifiers(oldStack, ref, entity, identifier);
+							Multimap<RegistryEntry<EntityAttribute>, EntityAttributeModifier> map = TrinketModifiers.get(oldStack, ref, entity);
 							Multimap<String, EntityAttributeModifier> slotMap = HashMultimap.create();
 							Set<RegistryEntry<EntityAttribute>> toRemove = Sets.newHashSet();
 							for (RegistryEntry<EntityAttribute> attr : map.keySet()) {
@@ -189,8 +190,7 @@ public abstract class LivingEntityMixin extends Entity {
 						}
 
 						if (!newStack.isEmpty()) {
-							Trinket trinket = TrinketsApi.getTrinket(newStack.getItem());
-							Multimap<RegistryEntry<EntityAttribute>, EntityAttributeModifier> map = trinket.getModifiers(newStack, ref, entity, identifier);
+							Multimap<RegistryEntry<EntityAttribute>, EntityAttributeModifier> map = TrinketModifiers.get(newStack, ref, entity);
 							Multimap<String, EntityAttributeModifier> slotMap = HashMultimap.create();
 							Set<RegistryEntry<EntityAttribute>> toRemove = Sets.newHashSet();
 							for (RegistryEntry<EntityAttribute> attr : map.keySet()) {
